Ngetn test function added
Nget n uses get with NEXT operator to get a sequence of up to n records
starting with either a random key or it uses key = 0 as a starting
value. The function takes 4 parameters:
sequence_size size_of_key exact_size
if exact_size is 0 then the string/byte key is up to size_of_key len,
otherwise it is exactly that size, if size_of_key is zero the starting
key for the sequence has value = 0.