Sample mdhimtst decks
Basic is a simple set of test comamnds to establish basic functionality.
Large is for large values and keys should only be used with key type of
strings and byte databases.
Next test the basic get operations.
Rand is used to generate large put runs of random keys and  values,
which can be of fixed or variable length.