Flag for append and NGetOp added
Test function for NGetOP (get next N) records added, and flag for
append DB_option. This flag turns on append mode. When records are
"put" if the key is already present the new value is appended to the
old value as opposed to the default mode of replacing the old value by
the new value.